hardware/server/requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
Pillow>=10.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
opencv-python>=4.8.0
python-dotenv>=1.0.0
gunicorn>=21.0.0
